feat: extract national/regional toggle from spatial extent field and hide it on the form
