+++ libs/feature/editor/src/lib/components/record-form/form-field/form-field-spatial-toggle/form-field-spatial-toggle.component.html
@@ -1,1 +1,5 @@
-<p>form-field-spatial-toggle works!</p>
+<gn-ui-switch-toggle
+  [options]="switchToggleOptions$ | async"
+  (selectedValue)="onSpatialScopeChange($event)"
+  extraClasses="grow text-sm"
+></gn-ui-switch-toggle>
